Add email addresses to the schools table
Email addresses are probably omitted from the main GIAS export because
of spamming, but they are available via FOI requests. As they are
out in the open and often needed by user resarchers to contact schools,
this modification adds them to the main schools table under the column
'email_address'.

The email list was sourced from this FOI request, all columns other than
the URN and MailEmail have been removed as much of it overlaps with the
main GIAS dataset.

https://www.whatdotheyknow.com/request/uk_schools_contact_data