Fix the urban/rural classification importer
The classifications appear to have changed and the changes have made the
cause problems in addition to the data uglier and less-consistent:

* prefixing some with (England/Wales) pushes the enum text length over
  64 characters and PostgreSQL won't allow it
* the capitalisation of the 'same' value isn't consistent, sometimes all
  words are capitalised and others just the first

Now we are uppercasing all of them and removing country-specific
prefixes. We know the exact point where the school lies and it's LA, so
it should be trivial to work out to which classification it belongs.